分组管理 增加双链表
